Switch to using QueryPerformaceCounter for more accurate FPS. And removes all the remaining drop frames code.

+++ src/main.h
@@ -53,8 +53,8 @@ BOOL IsWindowsXp();
 
 bool GameHandlesClose;
 int DrawFPS;
-DWORD TargetFPS;
-DWORD TargetFrameLen;
+double TargetFPS;
+double TargetFrameLen;
 bool SingleProcAffinity;
 int SwapInterval;
